+++ xarray/__init__.py
@@ -5,6 +5,7 @@ from .core.extensions import (register_dataarray_accessor,
 from .core.variable import Variable, Coordinate
 from .core.dataset import Dataset
 from .core.dataarray import DataArray
+from .core.merge import merge, MergeError
 from .core.options import set_options
 
 from .backends.api import open_dataset, open_mfdataset, save_mfdataset