+++ xarray/core/dataset.py
@@ -14,13 +14,13 @@ from . import indexing
 from . import alignment
 from . import formatting
 from .. import conventions
-from .alignment import align, align_variables
+from .alignment import align
 from .coordinates import DatasetCoordinates, Indexes
 from .common import ImplementsDatasetReduce, BaseDataObject
-from .merge import merge_datasets, expand_variables
-from .utils import Frozen, SortedKeysDict, ChainMap, maybe_wrap_array, hashable
-from .variable import (as_variable, Variable, Coordinate, broadcast_variables,
-                       default_index_coordinate)
+from .merge import (dataset_update_method, dataset_merge_method,
+                    merge_data_and_coords)
+from .utils import Frozen, SortedKeysDict, maybe_wrap_array, hashable
+from .variable import (Variable, as_variable, Coordinate, broadcast_variables)
 from .pycompat import (iteritems, basestring, OrderedDict,
                        dask_array_type)
 from .combine import concat
@@ -63,19 +63,19 @@ def _get_virtual_variable(variables, key):
     return ref_name, var_name, Variable(ref_var.dims, data)
 
 
-def _calculate_dims(variables):
+def calculate_dimensions(variables):
     """Calculate the dimensions corresponding to a set of variables.
 
     Returns dictionary mapping from dimension names to sizes. Raises ValueError
     if any of the dimension sizes conflict.
     """
-    dims = {}
+    dims = OrderedDict()
     last_used = {}
     scalar_vars = set(k for k, v in iteritems(variables) if not v.dims)
     for k, var in iteritems(variables):
         for dim, size in zip(var.dims, var.shape):
             if dim in scalar_vars:
-                raise ValueError('dimension %s already exists as a scalar '
+                raise ValueError('dimension %r already exists as a scalar '
                                  'variable' % dim)
             if dim not in dims:
                 dims[dim] = size
@@ -105,7 +105,7 @@ def as_dataset(obj):
     return obj
 
 
-class DataVariables(Mapping):
+class DataVariables(Mapping, formatting.ReprMixin):
     def __init__(self, dataset):
         self._dataset = dataset
 
@@ -126,7 +126,7 @@ class DataVariables(Mapping):
         else:
             raise KeyError(key)
 
-    def __repr__(self):
+    def __unicode__(self):
         return formatting.vars_repr(self)
 
 
@@ -140,7 +140,8 @@ class _LocIndexer(object):
         return self.dataset.sel(**key)
 
 
-class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject):
+class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject,
+              formatting.ReprMixin):
     """A multi-dimensional, in memory, array database.
 
     A dataset resembles an in-memory representation of a NetCDF file, and
@@ -198,13 +199,15 @@ class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject):
         if kwargs:
             if 'variables' in kwargs:
                 data_vars = kwargs.pop('variables')
-                warnings.warn('Variables kwarg is deprecated. Use data_vars', stacklevel=2)
+                warnings.warn('`variables` kwarg is deprecated. Use '
+                              '`data_vars` instead.', stacklevel=2)
             if kwargs:
-                raise TypeError('{0} are not valid kwargs'.format(kwargs.keys()))
+                raise TypeError(
+                    '{0} are not valid kwargs'.format(kwargs.keys()))
         if data_vars is None:
             data_vars = {}
         if coords is None:
-            coords = set()
+            coords = {}
         if data_vars is not None or coords is not None:
             self._set_init_vars_and_dims(data_vars, coords, compat)
         if attrs is not None:
@@ -218,55 +221,20 @@ class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject):
             if dim not in self._variables:
                 self._variables[dim] = default_index_coordinate(dim, size)
 
-    def _update_vars_and_coords(self, new_variables, new_coord_names=None,
-                                needs_copy=True, check_coord_names=True):
-        """Add a dictionary of new variables to this dataset.
-
-        Raises a ValueError if any dimensions have conflicting lengths in the
-        new dataset. Otherwise will update this dataset's _variables and
-        _dims attributes in-place.
-
-        Set `needs_copy=False` only if this dataset is brand-new and hence
-        can be thrown away if this method fails.
-        """
-        if new_coord_names is None:
-            new_coord_names = {}
-        # default to creating another copy of variables so can unroll if we end
-        # up with inconsistent dimensions
-        variables = self._variables.copy() if needs_copy else self._variables
-
-        if check_coord_names:
-            _assert_empty([k for k in self.data_vars if k in new_coord_names],
-                          'coordinates with these names already exist as '
-                          'variables: %s')
-
-        variables.update(new_variables)
-        dims = _calculate_dims(variables)
-        # all checks are complete: it's safe to update
-        self._variables = variables
-        self._dims = dims
-        self._add_missing_coords_inplace()
-        self._coord_names.update(new_coord_names)
-
-    def _set_init_vars_and_dims(self, vars, coords, compat):
+    def _set_init_vars_and_dims(self, data_vars, coords, compat):
         """Set the initial value of Dataset variables and dimensions
         """
-        _assert_empty([k for k in vars if k in coords],
-                      'redundant variables and coordinates: %s')
+        both_data_and_coords = [k for k in data_vars if k in coords]
+        if both_data_and_coords:
+            raise ValueError('variables %r are found in both data_vars and '
+                             'coords' % both_data_and_coords)
 
-        # Need to coerce Series to a dict as it doesn't implement __iter__ like a mapping
-        # GH470
-        vars = dict(vars) if isinstance(vars, pd.Series) else vars
+        variables, coord_names, dims = merge_data_and_coords(
+            data_vars, coords, compat=compat)
 
-        variables = ChainMap(vars, coords)
-
-        aligned = align_variables(variables)
-        new_variables, new_coord_names = expand_variables(aligned,
-                                                          compat=compat)
-
-        new_coord_names.update(coords)
-        self._update_vars_and_coords(new_variables, new_coord_names,
-                                     needs_copy=False, check_coord_names=False)
+        self._variables = variables
+        self._coord_names = coord_names
+        self._dims = dims
 
     @classmethod
     def load_store(cls, store, decoder=None):
@@ -365,7 +333,7 @@ class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject):
         return self.load()
 
     @classmethod
-    def _construct_direct(cls, variables, coord_names, dims, attrs,
+    def _construct_direct(cls, variables, coord_names, dims=None, attrs=None,
                           file_obj=None):
         """Shortcut around __init__ for internal use when we want to skip
         costly validation
@@ -381,12 +349,17 @@ class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject):
 
     __default_attrs = object()
 
-    def _replace_vars_and_dims(self, variables, coord_names=None,
+    @classmethod
+    def _from_vars_and_coord_names(cls, variables, coord_names):
+        dims = calculate_dimensions(variables)
+        return cls._construct_direct(variables, coord_names, dims)
+
+    def _replace_vars_and_dims(self, variables, coord_names=None, dims=None,
                                attrs=__default_attrs, inplace=False):
         """Fastpath constructor for internal use.
 
-        Preserves coord names and attributes; dimensions are recalculated from
-        the supplied variables.
+        Preserves coord names and attributes. If not provided explicitly,
+        dimensions are recalculated from the supplied variables.
 
         The arguments are *not* copied when placed on the new dataset. It is up
         to the caller to ensure that they have the right type and are not used
@@ -402,7 +375,8 @@ class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject):
         -------
         new : Dataset
         """
-        dims = _calculate_dims(variables)
+        if dims is None:
+            dims = calculate_dimensions(variables)
         if inplace:
             self._dims = dims
             self._variables = variables
@@ -417,6 +391,23 @@ class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject):
             if attrs is self.__default_attrs:
                 attrs = self._attrs_copy()
             obj = self._construct_direct(variables, coord_names, dims, attrs)
+        return obj
+
+    def _replace_indexes(self, indexes):
+        if not len(indexes):
+            return self
+        variables = self._variables.copy()
+        for name, idx in indexes.items():
+            variables[name] = Coordinate(name, idx)
+        obj = self._replace_vars_and_dims(variables)
+
+        # switch from dimension to level names, if necessary
+        dim_names = {}
+        for dim, idx in indexes.items():
+            if not isinstance(idx, pd.MultiIndex) and idx.name != dim:
+                dim_names[dim] = idx.name
+        if dim_names:
+            obj = obj.rename(dim_names)
         return obj
 
     def copy(self, deep=False):
@@ -635,7 +626,7 @@ class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject):
     def indexes(self):
         """OrderedDict of pandas.Index objects used for label based indexing
         """
-        return Indexes(self)
+        return Indexes(self._variables, self._dims)
 
     @property
     def coords(self):
@@ -793,7 +784,7 @@ class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject):
     dump = utils.function_alias(to_netcdf, 'dump')
     dumps = utils.function_alias(to_netcdf, 'dumps')
 
-    def __repr__(self):
+    def __unicode__(self):
         return formatting.dataset_repr(self)
 
     @property
@@ -944,7 +935,7 @@ class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject):
             Method to use for inexact matches (requires pandas>=0.16):
 
             * None (default): only exact matches
-            * pad / ffill: propgate last valid index value forward
+            * pad / ffill: propagate last valid index value forward
             * backfill / bfill: propagate next valid index value backward
             * nearest: use nearest valid index value
         tolerance : optional
@@ -954,7 +945,9 @@ class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject):
             Requires pandas>=0.17.
         **indexers : {dim: indexer, ...}
             Keyword arguments with names matching dimensions and values given
-            by scalars, slices or arrays of tick labels.
+            by scalars, slices or arrays of tick labels. For dimensions with
+            multi-index, the indexer may also be a dict-like object with keys
+            matching index level names.
 
         Returns
         -------
@@ -972,8 +965,10 @@ class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject):
         Dataset.isel_points
         DataArray.sel
         """
-        return self.isel(**indexing.remap_label_indexers(
-            self, indexers, method=method, tolerance=tolerance))
+        pos_indexers, new_indexes = indexing.remap_label_indexers(
+            self, indexers, method=method, tolerance=tolerance
+        )
+        return self.isel(**pos_indexers)._replace_indexes(new_indexes)
 
     def isel_points(self, dim='points', **indexers):
         """Returns a new dataset with each array indexed pointwise along the
@@ -1012,6 +1007,8 @@ class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject):
         Dataset.sel_points
         DataArray.isel_points
         """
+        from .dataarray import DataArray
+
         indexer_dims = set(indexers)
 
         def relevant_keys(mapping):
@@ -1050,10 +1047,9 @@ class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject):
                 # dim already exists
                 raise ValueError('Existing dimensions are not valid choices '
                                  'for the dim argument in sel_points')
-        else:
-            # try to cast dim to DataArray with name = points
-            from .dataarray import DataArray
-            dim = DataArray(dim, dims='points', name='points')
+
+        if not utils.is_scalar(dim) and not isinstance(dim, DataArray):
+            dim = as_variable(dim, name='points')
 
         # TODO: This would be sped up with vectorized indexing. This will
         # require dask to support pointwise indexing as well.
@@ -1114,8 +1110,9 @@ class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject):
         Dataset.isel_points
         DataArray.sel_points
         """
-        pos_indexers = indexing.remap_label_indexers(
-            self, indexers, method=method, tolerance=tolerance)
+        pos_indexers, _ = indexing.remap_label_indexers(
+            self, indexers, method=method, tolerance=tolerance
+        )
         return self.isel_points(dim=dim, **pos_indexers)
 
     def reindex_like(self, other, method=None, tolerance=None, copy=True):
@@ -1396,9 +1393,6 @@ class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject):
         obj = self.reindex(copy=False, **{dim: full_idx})
 
         new_dim_names = index.names
-        if any(name is None for name in new_dim_names):
-            raise ValueError('cannot unstack dimension with unnamed levels')
-
         new_dim_sizes = [lev.size for lev in index.levels]
 
         variables = OrderedDict()
@@ -1439,8 +1433,10 @@ class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject):
             If any dimensions would have inconsistent sizes in the updated
             dataset.
         """
-        return self.merge(
-            other, inplace=inplace, overwrite_vars=list(other), join='left')
+        variables, coord_names, dims = dataset_update_method(self, other)
+
+        return self._replace_vars_and_dims(variables, coord_names, dims,
+                                           inplace=inplace)
 
     def merge(self, other, inplace=False, overwrite_vars=set(),
               compat='broadcast_equals', join='outer'):
@@ -1488,11 +1484,11 @@ class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject):
         ValueError
             If any variables conflict (see ``compat``).
         """
-        replace_vars, new_coord_names = merge_datasets(
+        variables, coord_names, dims = dataset_merge_method(
             self, other, overwrite_vars, compat=compat, join=join)
-        obj = self if inplace else self.copy()
-        obj._update_vars_and_coords(replace_vars, new_coord_names)
-        return obj
+
+        return self._replace_vars_and_dims(variables, coord_names, dims,
+                                           inplace=inplace)
 
     def _assert_all_in_dataset(self, names, virtual_okay=False):
         bad_names = set(names) - set(self._variables)
@@ -1510,8 +1506,8 @@ class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject):
 
         Parameters
         ----------
-        labels : str
-            Names of variables or index labels to drop.
+        labels : scalar or list of scalars
+            Name(s) of variables or index labels to drop.
         dim : None or str, optional
             Dimension along which to drop index labels. By default (if
             ``dim is None``), drops variables rather than index labels.
@@ -1756,7 +1752,7 @@ class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject):
 
         coord_names = set(k for k in self.coords if k in variables)
         attrs = self.attrs if keep_attrs else None
-        return self._replace_vars_and_dims(variables, coord_names, attrs)
+        return self._replace_vars_and_dims(variables, coord_names, attrs=attrs)
 
     def apply(self, func, keep_attrs=False, args=(), **kwargs):
         """Apply a function over the data variables in this dataset.
@@ -1946,13 +1942,14 @@ class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject):
             if isinstance(other, groupby.GroupBy):
                 raise TypeError('in-place operations between a Dataset and '
                                 'a grouped object are not permitted')
+            # we don't actually modify arrays in-place with in-place Dataset
+            # arithmetic -- this lets us automatically align things
             if hasattr(other, 'indexes'):
                 other = other.reindex_like(self, copy=False)
-            # we don't want to actually modify arrays in-place
             g = ops.inplace_to_noninplace_op(f)
             ds = self._calculate_binary_op(g, other, inplace=True)
             self._replace_vars_and_dims(ds._variables, ds._coord_names,
-                                        ds._attrs, inplace=True)
+                                        attrs=ds._attrs, inplace=True)
             return self
         return func