+++ xarray/test/test_plot.py
@@ -349,7 +349,7 @@ class TestDetermineCmapParams(TestCase):
         self.assertEqual(cmap_params['vmax'], 0.9)
         self.assertEqual(cmap_params['cmap'], "RdBu_r")
 
-        # Setting vmin or vmax alone will force symetric bounds around center
+        # Setting vmin or vmax alone will force symmetric bounds around center
         cmap_params = _determine_cmap_params(neg, vmin=-0.1)
         self.assertEqual(cmap_params['vmin'], -0.1)
         self.assertEqual(cmap_params['vmax'], 0.1)
@@ -414,7 +414,7 @@ class TestDiscreteColorMap(TestCase):
             if filled:
                 self.assertEqual(ncmap.colorbar_extend, extend)
             else:
-                self.assertEqual(ncmap.colorbar_extend, 'neither')
+                self.assertEqual(ncmap.colorbar_extend, 'max')
 
     def test_discrete_colormap_list_of_levels(self):
         for extend, levels in [('max', [-1, 2, 4, 8, 10]),
@@ -429,7 +429,7 @@ class TestDiscreteColorMap(TestCase):
                 if kind != 'contour':
                     self.assertEqual(extend, primitive.cmap.colorbar_extend)
                 else:
-                    self.assertEqual('neither', primitive.cmap.colorbar_extend)
+                    self.assertEqual('max', primitive.cmap.colorbar_extend)
                 self.assertEqual(len(levels) - 1, len(primitive.cmap.colors))
 
     def test_discrete_colormap_int_levels(self):
@@ -454,7 +454,7 @@ class TestDiscreteColorMap(TestCase):
                 if kind != 'contour':
                     self.assertEqual(extend, primitive.cmap.colorbar_extend)
                 else:
-                    self.assertEqual('neither', primitive.cmap.colorbar_extend)
+                    self.assertEqual('max', primitive.cmap.colorbar_extend)
                 self.assertGreaterEqual(levels, len(primitive.cmap.colors))
 
     def test_discrete_colormap_list_levels_and_vmin_or_vmax(self):
@@ -593,7 +593,7 @@ class Common2dMixin:
         ax = plt.gca()
         self.assertEqual('x', ax.get_xlabel())
         self.assertEqual('newy', ax.get_ylabel())
-        # ax limits might change bewteen plotfuncs
+        # ax limits might change between plotfuncs
         # simply ensure that these high coords were passed over
         self.assertTrue(np.min(ax.get_ylim()) > 100.)
 
@@ -607,7 +607,7 @@ class Common2dMixin:
         ax = plt.gca()
         self.assertEqual('newy', ax.get_xlabel())
         self.assertEqual('x', ax.get_ylabel())
-        # ax limits might change bewteen plotfuncs
+        # ax limits might change between plotfuncs
         # simply ensure that these high coords were passed over
         self.assertTrue(np.min(ax.get_xlim()) > 100.)
 
@@ -618,9 +618,9 @@ class Common2dMixin:
         title = plt.gca().get_title()
         self.assertTrue('c = 1, d = foo' == title or 'd = foo, c = 1' == title)
 
-    def test_colorbar_label(self):
+    def test_colorbar_default_label(self):
         self.darray.name = 'testvar'
-        self.plotmethod()
+        self.plotmethod(add_colorbar=True)
         self.assertIn(self.darray.name, text_in_fig())
 
     def test_no_labels(self):
@@ -629,6 +629,43 @@ class Common2dMixin:
         alltxt = text_in_fig()
         for string in ['x', 'y', 'testvar']:
             self.assertNotIn(string, alltxt)
+
+    def test_colorbar_kwargs(self):
+        # replace label
+        self.darray.name = 'testvar'
+        self.plotmethod(add_colorbar=True, cbar_kwargs={'label':'MyLabel'})
+        alltxt = text_in_fig()
+        self.assertIn('MyLabel', alltxt)
+        self.assertNotIn('testvar', alltxt)
+        # you can use mapping types as well
+        self.plotmethod(add_colorbar=True, cbar_kwargs=(('label', 'MyLabel'),))
+        alltxt = text_in_fig()
+        self.assertIn('MyLabel', alltxt)
+        self.assertNotIn('testvar', alltxt)
+        # change cbar ax
+        fig, (ax, cax) = plt.subplots(1, 2)
+        self.plotmethod(ax=ax, cbar_ax=cax, add_colorbar=True,
+                        cbar_kwargs={'label':'MyBar'})
+        self.assertTrue(ax.has_data())
+        self.assertTrue(cax.has_data())
+        alltxt = text_in_fig()
+        self.assertIn('MyBar', alltxt)
+        self.assertNotIn('testvar', alltxt)
+        # note that there are two ways to achieve this
+        fig, (ax, cax) = plt.subplots(1, 2)
+        self.plotmethod(ax=ax, add_colorbar=True,
+                        cbar_kwargs={'label':'MyBar', 'cax':cax})
+        self.assertTrue(ax.has_data())
+        self.assertTrue(cax.has_data())
+        alltxt = text_in_fig()
+        self.assertIn('MyBar', alltxt)
+        self.assertNotIn('testvar', alltxt)
+        # see that no colorbar is respected
+        self.plotmethod(add_colorbar=False)
+        self.assertNotIn('testvar', text_in_fig())
+        # check that error is raised
+        self.assertRaises(ValueError, self.plotmethod,
+                          add_colorbar=False, cbar_kwargs= {'label':'label'})
 
     def test_verbose_facetgrid(self):
         a = easy_array((10, 15, 3))
@@ -747,7 +784,7 @@ class TestContour(Common2dMixin, PlotTestCase):
 
     def test_colors(self):
         # matplotlib cmap.colors gives an rgbA ndarray
-        # when seaborn is used, instead we get an rgb tuble
+        # when seaborn is used, instead we get an rgb tuple
         def _color_as_tuple(c):
             return tuple(c[:3])
         artist = self.plotmethod(colors='k')
@@ -758,6 +795,19 @@ class TestContour(Common2dMixin, PlotTestCase):
         artist = self.plotmethod(colors=['k', 'b'])
         self.assertEqual(
             _color_as_tuple(artist.cmap.colors[1]),
+            (0.0, 0.0, 1.0))
+
+        artist = self.darray.plot.contour(levels=[-0.5, 0., 0.5, 1.],
+                                          colors=['k', 'r', 'w', 'b'])
+        self.assertEqual(
+            _color_as_tuple(artist.cmap.colors[1]),
+            (1.0, 0.0, 0.0))
+        self.assertEqual(
+            _color_as_tuple(artist.cmap.colors[2]),
+            (1.0, 1.0, 1.0))
+        # the last color is now under "over"
+        self.assertEqual(
+             _color_as_tuple(artist.cmap._rgba_over),
             (0.0, 0.0, 1.0))
 
     def test_cmap_and_color_both(self):
@@ -774,6 +824,12 @@ class TestContour(Common2dMixin, PlotTestCase):
         ax = plt.gca()
         self.assertEqual('x2d', ax.get_xlabel())
         self.assertEqual('y2d', ax.get_ylabel())
+
+    def test_single_level(self):
+        # this used to raise an error, but not anymore since
+        # add_colorbar defaults to false
+        self.plotmethod(levels=[0.1])
+        self.plotmethod(levels=1)
 
 
 class TestPcolormesh(Common2dMixin, PlotTestCase):